requirements.txt
sentence-transformers>=2.2.0
chromadb>=0.4.0
neo4j>=5.0.0
tqdm>=4.66.0
torch>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0

